Update init_package_loading.lua
on game start, there was an error "The current directory is not inside the noita-mp". fixed it, 
io.popen("cd"):read("*a") were return Noita directory, not mod
also don't think we need to check for correct directory, it can't be uncorrect
now it works(how i think) as it supposed to be. 

P.S: there are still error about libs, if i understood correctly, you want to move all libs from \lua_modules to \files\lib\external\ , so i didn't touch it (kinda strange that it's already defined in mod, but still not moved)